minor test refactor
